@@ -2,18 +2,22 @@
   "name": "web",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:9709",
   "dependencies": {
-    "@types/jest": "24.0.21",
-    "@types/node": "12.12.5",
+    "@material-ui/core": "^4.6.1",
+    "@material-ui/icons": "^4.5.1",
+    "@types/jest": "^24.0.22",
+    "@types/node": "^12.12.7",
     "@types/react": "16.9.11",
-    "@types/react-dom": "16.9.3",
+    "@types/react-dom": "^16.9.4",
+    "material-table": "^1.54.1",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
-    "react-three-fiber": "^3.0.8",
+    "react-three-fiber": "^3.0.13",
     "three": "^0.110.0",
-    "typescript": "3.6.4"
+    "typescript": "^3.7.2"
   },
   "scripts": {
     "start": "react-scripts start",
