@@ -1,1 +1,3 @@
-export const EQUIL_ENDPOINT="/api/equil";
+export const EQUIL_ENDPOINT = "/api/equil";
+export const DATA_ENDPOINT = "/api/datas";
+export const SEARCH_ENDPOINT = "/api/search";
